backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.28.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
bcrypt>=4.0.0
requests>=2.31.0
